refactor setup_metadata_controller to be spell_type specific
