+++ conjureup/consts.py
@@ -20,6 +20,11 @@ ALLOWED_CONSTRAINTS = [
 ]
 
 
+class spell_types:
+    JUJU = 'juju'
+    SNAP = 'snap'
+
+
 class cloud_types:
     AWS = 'ec2'
     MAAS = 'maas'
