+++ conjureup/controllers/__init__.py
@@ -38,8 +38,85 @@ c.finish()
 from functools import lru_cache
 from importlib import import_module
 
-from conjureup import events
+from conjureup import events, consts
 from conjureup.app_config import app
+import yaml
+from conjureup import charm
+from conjureup.utils import slurp
+from conjureup.bundle import Bundle
+from pathlib import Path
+from itertools import chain
+
+
+def setup_metadata_controller():
+    """ Load metadata controller based on spell_type
+    """
+    if app.metadata.spell_type == consts.spell_types.SNAP:
+        return _setup_snap_metadata_controller()
+
+    # This is the typical default for now
+    return _setup_juju_metadata_controller()
+
+
+def _setup_snap_metadata_controller():
+    """ Sets metadata for a snap spell
+    """
+    return
+
+
+def _setup_juju_metadata_controller():
+    """ Pulls in a local bundle or via charmstore api and sets up our
+    controller. You can also further customize the bundle by providing a local
+    bundle-custom.yaml that will be deep merged over whatever bundle is
+    referenced. """
+    spell_dir = Path(app.config['spell-dir'])
+    bundle_filename = spell_dir / 'bundle.yaml'
+    bundle_custom_filename = spell_dir / 'bundle-custom.yaml'
+    if bundle_filename.exists():
+        # Load bundle data early so we can merge any additional charm options
+        bundle_data = Bundle(yaml.load(bundle_filename.read_text()))
+    else:
+        bundle_name = app.metadata.bundle_name
+        if bundle_name is None:
+            raise Exception(
+                "Could not determine a bundle to download, please make sure "
+                "the spell contains a 'bundle-name' field."
+            )
+        bundle_channel = app.conjurefile['channel']
+
+        app.log.debug("Pulling bundle for {} from channel: {}".format(
+            bundle_name, bundle_channel))
+        bundle_data = Bundle(charm.get_bundle(bundle_name, bundle_channel))
+
+    if bundle_custom_filename.exists():
+        bundle_custom = yaml.load(slurp(bundle_custom_filename))
+        bundle_data.apply(bundle_custom)
+
+    for name in app.selected_addons:
+        addon = app.addons[name]
+        bundle_data.apply(addon.bundle)
+
+    steps = list(chain(app.steps,
+                       chain.from_iterable(app.addons[addon].steps
+                                           for addon in app.selected_addons)))
+    for step in steps:
+        if not (step.bundle_add or step.bundle_remove):
+            continue
+        if step.bundle_remove:
+            fragment = yaml.safe_load(step.bundle_remove.read_text())
+            bundle_data.subtract(fragment)
+        if step.bundle_add:
+            fragment = yaml.safe_load(step.bundle_add.read_text())
+            bundle_data.apply(fragment)
+
+    if app.conjurefile['bundle-remove']:
+        fragment = yaml.safe_load(app.conjurefile['bundle-remove'].read_text())
+        bundle_data.subtract(fragment)
+    if app.conjurefile['bundle-add']:
+        fragment = yaml.safe_load(app.conjurefile['bundle-add'].read_text())
+        bundle_data.apply(fragment)
+
+    app.current_bundle = bundle_data
 
 
 @lru_cache(maxsize=None)
